MS-205 Reconfigure config-sync module to sync.config
