+++ feature/login-check/build.gradle.kts
@@ -12,7 +12,7 @@ dependencies {
 
     implementation(project(":infra:orchestrator-data"))
     implementation(project(":infra:config-store"))
-    implementation(project(":infra:config-sync"))
+    implementation(project(":infra:sync"))
     implementation(project(":infra:events"))
     implementation(project(":infra:event-sync"))
     implementation(project(":infra:images"))
