+++ id/build.gradle.kts
@@ -28,7 +28,7 @@ dependencies {
 
     implementation(project(":infra:core"))
     implementation(project(":infra:event-sync"))
-    implementation(project(":infra:config-sync"))
+    implementation(project(":infra:sync"))
     implementation(project(":infra:images"))
     implementation(project(":infra:auth-logic"))
     implementation(project(":infra:auth-store"))
