--- infra/config-sync/src/main/java/com/simprints/infra/config/sync/ProjectConfigurationScheduler.kt
@@ -1,10 +1,0 @@
-package com.simprints.infra.config.sync
-
-interface ProjectConfigurationScheduler {
-    fun scheduleProjectSync()
-    fun cancelProjectSync()
-
-    fun startDeviceSync()
-    fun scheduleDeviceSync()
-    fun cancelDeviceSync()
-}
