+++ infra/sync/.gitignore
