+++ infra/sync/consumer-rules.pro
