+++ infra/sync/src/main/AndroidManifest.xml
