+++ infra/sync/src/main/java/com/simprints/infra/sync/config/ProjectConfigurationScheduler.kt
@@ -1,0 +1,10 @@
+package com.simprints.infra.sync.config
+
+interface ProjectConfigurationScheduler {
+    fun scheduleProjectSync()
+    fun cancelProjectSync()
+
+    fun startDeviceSync()
+    fun scheduleDeviceSync()
+    fun cancelDeviceSync()
+}
